🐛 Fixed Offers erroring when be hydrated from the database
Bad data should never cause errors when reading from the database, we can never
be sure of the state of data in the database. This patch will ensure we don't
run the validations when hydrating offers from the database. But longer term we
need to decide on a sensible default so that we are dealing with valid entities